[permalink] Ensure that the permalink is correctly set #199 (#200)
It ensures that the permalink_lang attribute is set correctly and can be used when a permalink attribute is set at the site level.
It is, indeed, not enough to check if the permalink is undefined. The permalink might be set to empty. Moreover, we should check that the page has a lang attribute isn't empty as well.